Moved theme.dart to correct folder in patient app
