+++ patient/lib/presentation/assessments/assessment_screen.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:patient/presentation/services/theme.dart';
+import 'package:patient/core/theme/theme.dart';
 import 'package:patient/provider/assessment_provider.dart';
 import 'package:provider/provider.dart';
